Add scalar kernel args, atomic operations, and field reductions
Scalar kernel arguments: Python scalars (int/float) can now be passed
directly to kernels instead of wrapping in fields. All backends updated
with _is_field param tracking and distinct codegen paths (pointer vs value).

Atomic operations: pgc.atomic_add/min/max(field, index, value) supported
across all backends — LLVM (atomic_rmw + CAS for float min/max), MSL
(atomic_fetch_add + CAS), CUDA/HIP (atomicAdd/Min/Max). IR passes
(resolve, optimize) updated to handle IRAtomicOp.

Field reductions: field.sum(), field.min(), field.max() host-side methods.

Float scalars now infer as f32 (not f64) for GPU compatibility.

+++ tests/test_type_inference.py
@@ -34,7 +34,7 @@ def test_infer_scalar_types():
     ], [])
 
     types = infer_param_types(func, (x, 2.5, 10))
-    assert types == [f32, f64, i32]
+    assert types == [f32, f32, i32]
 
 
 def test_infer_wrong_arg_count():
